Changed column width multiplier
Changed from 5 to 7 and it is very nice looking, and lets the font changes be more open to customizing the experience.  My screen resolution is 1680x1050 so I am hoping that it will look good on other screen sizes. I will see how hard it might be, to make a formula for assigning percentages to each column width, independent of font and resolution.